# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portalu API",
    "version": "2.3.0"
  },
  "servers": [
    {
      "url": "https://portalu-production.up.railway.app",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/v1/auth/sign-up": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign Up",
        "operationId": "sign_up_api_v1_auth_sign_up_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify Email",
        "operationId": "verify_email_api_v1_auth_verify_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/sign-in": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign In",
        "operationId": "sign_in_api_v1_auth_sign_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sign-out": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign Out",
        "operationId": "sign_out_api_v1_auth_sign_out_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot Password",
        "operationId": "forgot_password_api_v1_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_api_v1_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_v1_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/sign-up": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Sign Up",
        "operationId": "sign_up_api_v1_citizen__org_slug___portal_slug__auth_sign_up_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenSignUpRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitizenSignUpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/verify-email": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Verify Email",
        "operationId": "verify_email_api_v1_citizen__org_slug___portal_slug__auth_verify_email_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenVerifyEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/sign-in": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Sign In",
        "operationId": "sign_in_api_v1_citizen__org_slug___portal_slug__auth_sign_in_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenSignInRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitizenAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/refresh": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Refresh",
        "operationId": "refresh_api_v1_citizen__org_slug___portal_slug__auth_refresh_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitizenAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/sign-out": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Sign Out",
        "operationId": "sign_out_api_v1_citizen__org_slug___portal_slug__auth_sign_out_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/forgot-password": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Forgot Password",
        "operationId": "forgot_password_api_v1_citizen__org_slug___portal_slug__auth_forgot_password_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/reset-password": {
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_api_v1_citizen__org_slug___portal_slug__auth_reset_password_post",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/auth/me": {
      "get": {
        "tags": [
          "citizen"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_citizen__org_slug___portal_slug__auth_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitizenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/request-types": {
      "get": {
        "tags": [
          "citizen"
        ],
        "summary": "List Request Types",
        "operationId": "list_request_types_api_v1_citizen__org_slug___portal_slug__request_types_get",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestTypeOut"
                  },
                  "title": "Response List Request Types Api V1 Citizen  Org Slug   Portal Slug  Request Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/requests": {
      "get": {
        "tags": [
          "citizen"
        ],
        "summary": "List My Requests",
        "operationId": "list_my_requests_api_v1_citizen__org_slug___portal_slug__requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CitizenRequestOut"
                  },
                  "title": "Response List My Requests Api V1 Citizen  Org Slug   Portal Slug  Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "citizen"
        ],
        "summary": "Create Request",
        "operationId": "create_request_api_v1_citizen__org_slug___portal_slug__requests_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitizenRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitizenRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/citizen/{org_slug}/{portal_slug}/requests/{request_id}": {
      "get": {
        "tags": [
          "citizen"
        ],
        "summary": "Get Request",
        "operationId": "get_request_api_v1_citizen__org_slug___portal_slug__requests__request_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "portal_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portal Slug"
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitizenRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ],
        "title": "AuthResponse"
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "SignInRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "SignInRequest"
      },
      "SignUpRequest": {
        "properties": {
          "org_name": {
            "type": "string",
            "title": "Org Name"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "org_name",
          "email",
          "password"
        ],
        "title": "SignUpRequest"
      },
      "SignUpResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "user",
          "message"
        ],
        "title": "SignUpResponse"
      },
...
      "VerifyEmailRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "VerifyEmailRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}